rename legacy functions, add messages sample
