+++ graylog-sdk-spring-boot-samples/build.gradle
@@ -11,6 +11,9 @@ dependencies {
     implementation group: 'org.projectlombok', name: 'lombok', version: '1.18.12'
     implementation group: 'com.joonsang.graylog', name: 'graylog-query-builder', version: '1.0.3'
 
+    implementation group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.8.0'
+    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.11'
+
     implementation project(':graylog-sdk-spring-boot-starter')
 
     annotationProcessor("org.projectlombok:lombok:1.18.12")
