+++ graylog-sdk-spring-boot-samples/src/main/java/com/joonsang/graylog/sdk/spring/samples/domain/LegacyHistograms.java
@@ -9,10 +9,10 @@ import java.util.List;
 
 @Getter
 @Setter
-public class Histograms {
+public class LegacyHistograms {
 
     @Builder
-    public Histograms(List<String> labels, List<Histogram> histograms) {
+    public LegacyHistograms(List<String> labels, List<Histogram> histograms) {
         this.labels = labels;
         this.histograms = histograms;
     }
